Refactor ModelLoader classes to integrate LoRAModelMixin
- Removed the __init__.py file from the gemma_lora module.
- Updated ModelLoader classes for Gemma, Llama, Phi-1, and Qwen 2.5 to inherit from LoRAModelMixin, centralizing LoRA adapter logic.
- Eliminated redundant load_model methods and configurations, enhancing code maintainability and reducing duplication.
- Introduced LoRAModelMixin to streamline LoRA adapter application and model loading processes across different model variants.
