--- gemma_lora/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: (c) 2026 Tenstorrent AI ULC
-#
-# SPDX-License-Identifier: Apache-2.0
-from .pytorch import ModelLoader, ModelVariant
